refactor: move IP address code to seperate file
