+++ proxmox/config_qemu_cloudinit.go
@@ -2,7 +2,6 @@ package proxmox
 
 import (
 	"errors"
-	"net"
 	"regexp"
 	"strconv"
 	"strings"
@@ -593,76 +592,6 @@ func (path CloudInitSnippetPath) Validate() error {
 	}
 	if !regexCloudInitSnippetPath_Path.MatchString(string(path)) {
 		return errors.New(CloudInitSnippetPath_Error_InvalidPath)
-	}
-	return nil
-}
-
-type IPv4Address string
-
-const IPv4Address_Error_Invalid = "ipv4Address is not a valid ipv6 address"
-
-func (ip IPv4Address) Validate() error {
-	if ip == "" {
-		return nil
-	}
-	if net.ParseIP(string(ip)) == nil {
-		return errors.New(IPv4Address_Error_Invalid)
-	}
-	if !isIPv4(string(ip)) {
-		return errors.New(IPv4Address_Error_Invalid)
-	}
-	return nil
-}
-
-type IPv4CIDR string
-
-const IPv4CIDR_Error_Invalid = "ipv4CIDR is not a valid ipv4 address"
-
-func (cidr IPv4CIDR) Validate() error {
-	if cidr == "" {
-		return nil
-	}
-	ip, _, err := net.ParseCIDR(string(cidr))
-	if err != nil {
-		return errors.New(IPv4CIDR_Error_Invalid)
-	}
-	if !isIPv4(ip.String()) {
-		return errors.New(IPv4CIDR_Error_Invalid)
-	}
-	return err
-}
-
-type IPv6Address string
-
-const IPv6Address_Error_Invalid = "ipv6Address is not a valid ipv6 address"
-
-func (ip IPv6Address) Validate() error {
-	if ip == "" {
-		return nil
-	}
-	if net.ParseIP(string(ip)) == nil {
-		return errors.New(IPv6Address_Error_Invalid)
-	}
-	if !isIPv6(string(ip)) {
-		return errors.New(IPv6Address_Error_Invalid)
-	}
-	return nil
-}
-
-type IPv6CIDR string
-
-const IPv6CIDR_Error_Invalid = "ipv6CIDR is not a valid ipv6 address"
-
-func (cidr IPv6CIDR) Validate() error {
-	if cidr == "" {
-		return nil
-	}
-	ip, _, err := net.ParseCIDR(string(cidr))
-	if err != nil {
-		return errors.New(IPv6CIDR_Error_Invalid)
-	}
-	if !isIPv6(ip.String()) {
-		return errors.New(IPv6CIDR_Error_Invalid)
 	}
 	return nil
 }
